ref:コントローラーとモデルのリファクタリング #471
