+++ app/controllers/design_tips_controller.rb
@@ -1,20 +1,17 @@
 class DesignTipsController < ApplicationController
   before_action :set_q, only: %i[index search]
   before_action :set_likes, only: %i[likes]
+  before_action :set_tag_list, only: %i[index search]
   skip_before_action :require_login
 
   def index
-    @search_design_tips = DesignTip.all
     @design_tips = @q.result(distinct: true).preload(:tags, :reviews)
-    @tag_list = DesignTip.tag_counts_on(:tags).most_used(20)
     @review = Review.new
     @user = User.new
   end
 
   def search
-    @search_design_tips = DesignTip.all.preload(:reviews)
     @results = @q.result
-    @tag_list = DesignTip.tag_counts_on(:tags).most_used(20)
   end
 
   def likes
@@ -25,6 +22,11 @@ class DesignTipsController < ApplicationController
 
   def set_q
     @q = DesignTip.ransack(params[:q])
+    @search_design_tips = DesignTip.all
+  end
+
+  def set_tag_list
+    @tag_list = DesignTip.tag_counts_on(:tags).most_used(20)
   end
 
   def set_likes
