+++ app/controllers/quizees_controller.rb
@@ -4,15 +4,6 @@ class QuizeesController < ApplicationController
   def index
     @quizees = Quiz.all
     @choices = Choice.all
-    check_answer
-  end
-
-  def check_answer
-    correct = params[:correct]
-    if correct == "true"
-      @result = "正解です！"
-    elsif correct == "false"
-      @result = "残念、不正解です..."
-    end
+    @result = Quiz.check_answer(params)
   end
 end
